refactor: componentizando dialog de deletar perfil
